chatbot/requirements.txt
Django
git+https://github.com/gunthercox/ChatterBot.git@master
chatterbot-corpus
SQLAlchemy
pint
mathparse
nltk
PyYAML
python-dateutil
